Implementacion de la funcion de aplicacion de credito, funcionan en paralelo las funciones anteriores y lo nuevo, se cambiaron desde tipos de datos hasta ajustes para que las mutaciones, cruces y cambios coexistieran con lo nuevo,pendiente el elitista

+++ src/main/java/com/example/demo/conversion/RealConverterService.java
@@ -10,7 +10,8 @@ public class RealConverterService {
 
     private static final Logger log = LoggerFactory.getLogger(RealConverterService.class);
 
-    public List<Double> toReal(List<Integer> decimals, double xmin, double xmax, int L) {
+    // 🚨 CAMBIO DE List<Integer> a List<Long>
+    public List<Double> toReal(List<Long> decimals, double xmin, double xmax, int L) {
         if (xmax <= xmin) {
             throw new IllegalArgumentException("xmax debe ser mayor que xmin");
         }
@@ -18,6 +19,7 @@ public class RealConverterService {
             throw new IllegalArgumentException("L debe ser un entero positivo");
         }
 
+        // L sigue siendo la longitud, pero Math.pow(2, L) se usará con cuidado si L > 52
         double maxDecimalValue = Math.pow(2, L) - 1;
         double range = xmax - xmin;
         double scaleFactor = range / maxDecimalValue;
@@ -25,6 +27,7 @@ public class RealConverterService {
         log.debug("Conversión binario → real: L={}, xmin={}, xmax={}, maxDecimal={}, scaleFactor={}",
                 L, xmin, xmax, maxDecimalValue, String.format("%.6f", scaleFactor));
 
+        // 🚨 CAMBIO DE Integer a Long
         return decimals.stream()
                 .map(v -> {
                     double real = xmin + (v * scaleFactor);
@@ -34,9 +37,12 @@ public class RealConverterService {
                 .collect(Collectors.toList());
     }
 
-    public Double toRealSingle(int decimal, double xmin, double xmax, int L) {
+    // 🚨 CAMBIO DE int a long
+    public Double toRealSingle(long decimal, double xmin, double xmax, int L) {
         double maxDecimalValue = Math.pow(2, L) - 1;
         double scaleFactor = (xmax - xmin) / maxDecimalValue;
+
+        // 🚨 USAR el valor long 'decimal'
         double real = xmin + (decimal * scaleFactor);
         log.trace("toRealSingle: {} → {}", decimal, String.format("%.6f", real));
         return real;
